Extracted catalog data processing into its own class.
