+++ src/Catalog/NuGet.Services.Metadata.Catalog.csproj
@@ -115,8 +115,10 @@
     <Compile Include="Helpers\PackageUtility.cs" />
     <Compile Include="Helpers\Retry.cs" />
     <Compile Include="ICatalogIndexProcessor.cs" />
+    <Compile Include="Icons\CatalogLeafDataProcessor.cs" />
     <Compile Include="Icons\ExternalIconCopyResult.cs" />
     <Compile Include="Icons\HttpResponseMessageProvider.cs" />
+    <Compile Include="Icons\ICatalogLeafDataProcessor.cs" />
     <Compile Include="Icons\IconProcessor.cs" />
     <Compile Include="Icons\IconsCollector.cs" />
     <Compile Include="Icons\IHttpResponseMessageProvider.cs" />
